refactor(web): migrate all type imports from lib/api to lib/types
